+++ Project/modules/platform/src/DynamixelProtocol.hpp
@@ -34,7 +34,7 @@ public:
     void DynamixelProtocol_init();
 };
 
-extern int executeWriteSingle(uint8_t servo_ID, uint8_t address, uint data);
+extern int executeWriteSingle(uint8_t servo_ID, uint8_t address, double data);
 extern int executeWriteMulti(uint8_t* servo_ID, uint8_t address, double** data, uint8_t count);
 extern int executeReadSingle(uint8_t servo_ID, uint8_t address, uint* data);
 extern int executeReadMulti(uint8_t* servo_ID, uint8_t address, uint* data, uint8_t count);
